render.js aside 베스트게시글 관련 함수 이동(20.11.09)
베스트게시글 관련 render함수 aside폴더 render.js 로 이동
