+++ TeamProject/templates/post.html
@@ -89,6 +89,7 @@
 <script type="module" src="/static/js/board/fetch.js"></script>
 <script type="module" src="/static/js/board/router.js"></script>
 <script type="module" src="/static/js/board/list/render.js"></script>
+<script type="module" src="/static/js/board/aside/render.js"></script>
 
 <script type="module" src="/static/js/auth.js"></script>
